+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New verb: `scan_templates`: scan a continuous waveform stream using one
 - New verb: `print_families`: print families to screen
+- New verb: `plot_slip`: plot cumulative slip for one or more families
 - `family_numbers` argument is no more mandatory (default: `all`)
   - argument added to `map_families` and `plot_timespans`
 - Config parameter `trace_average_from_normalized_traces` renamed to